requirements.txt
gradio==5.20.1
langchain==0.3.20
langchain-groq==0.2.5
langchain-ollama==0.2.3
langchain-openai==0.3.9
python-dotenv==1.0.1